EdGE issues were fixed
Summary:
This diff fixes T148

Edge based browsers seemed to have an issue with the pop-up blocker not being able to
open a blob once. Which caused the browser to catch the user in an infinite loop of asking
whether they wanted to open the PDF just once and then not opening it.

Test Plan: The widget now should display a file download dialog instead.

Reviewers: sebastian

Reviewed By: sebastian

Maniphest Tasks: T148

Differential Revision: https://phabricator.magic3w.com/D284